refactor (post) : 불필요한 추상화 삭제하여 코드 가독성 및 유지보수성 향상
