+++ src/main/java/com/backend/immilog/post/application/services/CommentUploadService.java
@@ -1,4 +1,4 @@
-package com.backend.immilog.post.application;
+package com.backend.immilog.post.application.services;
 
 import com.backend.immilog.post.exception.PostException;
 import com.backend.immilog.post.model.entities.Comment;
@@ -6,22 +6,22 @@ import com.backend.immilog.post.model.entities.Post;
 import com.backend.immilog.post.model.enums.ReferenceType;
 import com.backend.immilog.post.model.repositories.CommentRepository;
 import com.backend.immilog.post.model.repositories.PostRepository;
-import com.backend.immilog.post.model.services.CommentUploadService;
 import com.backend.immilog.post.presentation.request.CommentUploadRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import static com.backend.immilog.post.exception.PostErrorCode.POST_NOT_FOUND;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class CommentUploadServiceImpl implements CommentUploadService {
+public class CommentUploadService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
 
-    @Override
+    @Transactional
     public void uploadComment(
             Long userId,
             Long postSeq,
