--- src/main/java/com/backend/immilog/post/model/services/PostDeleteService.java
@@ -1,11 +1,0 @@
-package com.backend.immilog.post.model.services;
-
-import org.springframework.transaction.annotation.Transactional;
-
-public interface PostDeleteService {
-    @Transactional
-    void deletePost(
-            Long userId,
-            Long postSeq
-    );
-}
