--- src/main/java/com/backend/immilog/post/model/services/PostInquiryService.java
@@ -1,34 +1,0 @@
-package com.backend.immilog.post.model.services;
-
-import com.backend.immilog.post.model.dtos.PostDTO;
-import com.backend.immilog.post.model.enums.Categories;
-import com.backend.immilog.post.model.enums.Countries;
-import com.backend.immilog.post.model.enums.SortingMethods;
-import org.springframework.data.domain.Page;
-import org.springframework.transaction.annotation.Transactional;
-
-public interface PostInquiryService {
-    @Transactional(readOnly = true)
-    Page<PostDTO> getPosts(
-            Countries country,
-            SortingMethods sortingMethod,
-            String isPublic,
-            Categories category,
-            Integer page
-    );
-
-    @Transactional(readOnly = true)
-    PostDTO getPost(Long postSeq);
-
-    @Transactional(readOnly = true)
-    Page<PostDTO> searchKeyword(
-            String keyword,
-            Integer page
-    );
-
-    @Transactional(readOnly = true)
-    Page<PostDTO> getUserPosts(
-            Long userSeq,
-            Integer page
-    );
-}
